+++ src/main/java/database/innodb/page/FileHeader.java
@@ -1,4 +1,4 @@
-package database.page;
+package database.innodb.page;
 
 import java.io.Serializable;
