+++ src/test/java/database/innodb/BufferPoolTest.java
@@ -1,9 +1,9 @@
-package database;
+package database.innodb;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import database.page.PageFactory;
+import database.innodb.page.PageFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
