+++ src/test/java/database/innodb/page/PageManagerTest.java
@@ -1,4 +1,4 @@
-package database.page;
+package database.innodb.page;
 
 import static org.assertj.core.api.Assertions.assertThat;
